+++ ci-scripts/ui-e2e.sh
@@ -71,7 +71,7 @@ echo "starting skycoin node in background with http listener on $HOST"
                       -db-read-only=true \
                       -launch-browser=false \
                       -data-dir="$DATA_DIR" \
-                      -enable-api-set=ALL \
+                      -enable-api-set=READ,STATUS,WALLET,WALLET_SEED \
                       -wallet-dir="$WALLET_DIR" \
                       &
 SKYCOIN_PID=$!