+++ src/skycoin/skycoin.go
@@ -326,7 +326,7 @@ func (c *Coin) ConfigureDaemon() daemon.Config {
 	_, dc.Visor.EnableSeedAPI = c.config.Node.enabledAPISets[api.EndpointsWalletSeed]
 
 	_, dc.Gateway.EnableWalletAPI = c.config.Node.enabledAPISets[api.EndpointsWallet]
-	dc.Gateway.EnableSpendMethod = c.config.Node.EnableSpendEndpoint
+	_, dc.Gateway.EnableSpendMethod = c.config.Node.enabledAPISets[api.EndpointsDeprecatedWalletSpend]
 
 	// Initialize wallet default crypto type
 	cryptoType, err := wallet.CryptoTypeFromString(c.config.Node.WalletCryptoType)